# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v1
components:
  schemas:
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        total:
          type: number
          format: double
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        userName:
          type: string
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        name:
          type: string
        company:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        stock:
          type: integer
          format: int32
        isAccepted:
          type: boolean
        date:
          type: string
          format: date-time
        views_count:
          type: integer
          format: int32
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        imageBase64:
          type: string
    Review:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        username:
          type: string
        comment:
          type: string
        calification:
          type: integer
          format: int32
        creationDate:
          type: string
          format: date-time
    ReviewDTO:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        username:
          type: string
        comment:
          type: string
        calification:
          type: integer
          format: int32
        creationDate:
          type: string
          format: date
    UserCreationDTO:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: array
          items:
            type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        cart:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        cartPrice:
          type: number
          format: double
        historicalOrderPrices:
          type: array
          items:
            type: number
            format: double
        imageBase64:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: array
          items:
            type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderDTO"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        cart:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        cartPrice:
          type: number
          format: double
        historicalOrderPrices:
          type: array
          items:
            type: number
            format: double
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          $ref: "#/components/schemas/PageMetadata"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
servers:
  - url: https://localhost:8443
    description: Generated server url
paths:
...
  /api/v1/users:
    get:
      tags:
      - api-user-controller
      summary: Get all users
      description: Retrieve a list of all users in the system.
      operationId: getAllUsers
      responses:
        "200":
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
    put:
      tags:
      - api-user-controller
      summary: Update user profile
      description: Update the details of an existing user.
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreationDTO"
        required: true
      responses:
        "200":
          description: Profile updated successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - api-user-controller
      summary: Create a new user
      description: Register a new user in the system.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreationDTO"
        required: true
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Username already exists
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
  /api/v1/products/{id}:
    get:
      tags:
      - api-product-controller
      summary: Get product by ID
      description: Retrieve a product by its unique ID.
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      tags:
      - api-product-controller
      summary: Update a product
      description: Update the details of an existing product.
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - api-product-controller
      summary: Delete a product
      description: Delete a product by its unique ID.
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/products/accept:
    put:
      summary: Accept product
      tags:
      - api-product-controller
      operationId: acceptProduct
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/products:
    get:
      tags:
      - api-product-controller
      summary: Get products
      description: Retrieve products based on filters.
      operationId: getProducts
      parameters:
      - name: accepted
        in: query
        required: false
        schema:
          type: boolean
      - name: company
        in: query
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: No products found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModel"
    post:
      tags:
      - api-product-controller
      summary: Create a new product
      description: Create a new product with a given set of properties.
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/products/delete:
    delete:
      summary: Delete product
      tags:
      - api-product-controller
      operationId: deleteProduct_1
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK